backend/requirements-dev.txt
ruff==0.13.0
pre-commit==4.3.0
pytest==8.4.2
httpx==0.28.1
pytest-cov==7.0.0
pytest-asyncio==0.24.0
pytest-celery==1.2.1
debugpy==1.8.12
setuptools<81
aiosqlite==0.21.0
